backend/requirements.txt
boto3==1.34.162
google-cloud-vision==3.8.0
Pillow==10.4.0
